refactor: move generalMetadata component in view folder
