+++ src/ui/view/dataverse/component/generalMetadata/generalMetadata.tsx
@@ -1,9 +1,9 @@
 import type { FC } from 'react'
 import { memo } from 'react'
 import { useTranslation } from 'react-i18next'
-import type { IconName } from '@/component/icon/icon'
-import { Icon } from '@/component/icon/icon'
-import { activeLanguageWithDefault } from '@/languages/languages'
+import type { IconName } from '@/ui/component/icon/icon'
+import { Icon } from '@/ui/component/icon/icon'
+import { activeLanguageWithDefault } from '@/ui/languages/languages'
 import { convertToYearIfISODateTime } from '@/util/isoDateTime/isoDateTime'
 import './generalMetadata.scss'
 import './i18n/index'
