+++ src/ui/view/dataverse/component/generalMetadata/i18n/generalMetadata_de.json
