+++ src/ui/view/dataverse/component/generalMetadata/i18n/index.ts
@@ -1,8 +1,8 @@
 import generalMetadata_en from './generalMetadata_en.json'
 import generalMetadata_de from './generalMetadata_de.json'
 import generalMetadata_fr from './generalMetadata_fr.json'
-import type { I18nResource } from '@/i18n/utils'
-import { loadTranslations } from '@/i18n/utils'
+import type { I18nResource } from '@/ui/i18n/utils'
+import { loadTranslations } from '@/ui/i18n/utils'
 
 const i18nTranslations: I18nResource[] = [
   { lng: 'en', namespace: 'generalMetadata', resource: generalMetadata_en },
